♻️ Move Result type to @nosgestesclimat/core/lib/result
