+++ apps/site/src/components/user/UserEmailForm.tsx
@@ -11,7 +11,7 @@ import { useUpdateUserSettings } from '@/hooks/settings/useUpdateUserSettings'
 import { useClientTranslation } from '@/hooks/useClientTranslation'
 import { trackPosthogEvent } from '@/utils/analytics/trackEvent'
 import { formatEmail } from '@/utils/format/formatEmail'
-import { ok } from '@/utils/result'
+import { ok } from '@nosgestesclimat/core/lib/result'
 import { useRouter } from 'next/navigation'
 import { useCallback, useState, type ReactNode } from 'react'
 import type { SubmitHandler } from 'react-hook-form'
