+++ packages/core/src/lib/result.ts
